refactor(mobile): move reminders under shared/ to respect feature isolation
Feature modules must only import from shared/ (AGENTS.md), but
features/channels/message_actions.dart imported the reminder sheet and
service from features/reminders. Reminders are a cross-cutting
capability consumed by the channels feature, so relocate the module
wholesale to lib/shared/reminders (and its test to
test/shared/reminders) with no behavior change.
