+++ mobile/lib/features/channels/message_actions.dart
@@ -9,8 +9,8 @@ import '../../shared/theme/theme.dart';
 import '../../shared/custom_emoji/custom_emoji.dart';
 import '../../shared/custom_emoji/custom_emoji_provider.dart';
 import '../../shared/widgets/sheet_divider.dart';
-import '../reminders/remind_me_later_sheet.dart';
-import '../reminders/reminder_service.dart';
+import '../../shared/reminders/remind_me_later_sheet.dart';
+import '../../shared/reminders/reminder_service.dart';
 import 'channel_management_provider.dart';
 import 'emoji_picker.dart';
 import 'read_state/message_read_state.dart';
