+++ mobile/lib/shared/reminders/remind_me_later_sheet.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 
-import '../../shared/theme/theme.dart';
-import '../../shared/widgets/sheet_divider.dart';
+import '../theme/theme.dart';
+import '../widgets/sheet_divider.dart';
 import 'reminder_service.dart';
 import 'reminder_time_presets.dart';
